complete_authentication_system/flasker/requirements/testing.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
click
Flask-Migrate
redis
celery
python-dotenv
flask-login
flask-bcrypt
Flask-WTF
email_validator
PyJWT

pytest
pylint
